requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
loguru==0.6.0
multidict==6.0.2
python-environ==0.4.54
pytz==2022.1
SQLAlchemy==1.4.39
yarl==1.8.1
